bypass conversion if target exists

+++ tools/convert-models.py
@@ -23,6 +23,15 @@ def upload_blob(bucket_name, source_file_name, destination_blob_name):
             source_file_name, destination_blob_name
         )
     )
+    
+def blob_exist(bucket_name, destination_blob_name):
+  """Check if file already exists in s3"""
+    s3 = boto3.client('s3')
+    try:
+      s3.head_object(destination_blob_name)
+      return True
+    except s3.exceptions.NoSuchKey:
+      return False
 
 models = {
   'alexnet': 'https://download.pytorch.org/models/alexnet-owt-4df8aa71.pth',
@@ -86,16 +95,22 @@ models = {
 os.makedirs("models", exist_ok=True)
 
 for name, url in models.items():
-  m = load_state_dict_from_url(url, progress=False)
-  converted = {}
-  for nm, par in m.items():
-    converted.update([(nm, par.clone())])
   fpath = "models/" + name + ".pth"
-  torch.save(converted, fpath, _use_new_zipfile_serialization=True)
-  upload_blob(
-    "torch-pretrained-models",
-    fpath,
-    "models/vision/v2/" + fpath
-  )
-  # free disk space
-  os.remove(fpath)
+  
+  if blob_exist("torch-pretrained-models", f"models/vision/v2/{fpath}"):
+    print(f"--- file {name}.pth is already in the bucket. Bypassing conversion")
+    
+  else:
+    # download from url, convert and upload the converted weights
+    m = load_state_dict_from_url(url, progress=False)
+    converted = {}
+    for nm, par in m.items():
+      converted.update([(nm, par.clone())])
+    torch.save(converted, fpath, _use_new_zipfile_serialization=True)
+    upload_blob(
+      "torch-pretrained-models",
+      fpath,
+      "models/vision/v2/" + fpath
+    )
+    # free disk space
+    os.remove(fpath)
